+++ dhis-2/dhis-support/dhis-support-hibernate/src/main/java/org/hisp/dhis/hibernate/jsonb/type/ApiKeyAttributeJsonBinaryType.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright (c) 2004-2021, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.hibernate.jsonb.type;
+
+import java.util.List;
+
+import org.hisp.dhis.security.apikey.ApiTokenAttribute;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+
+public class ApiKeyAttributeJsonBinaryType extends JsonBinaryType
+{
+    public ApiKeyAttributeJsonBinaryType()
+    {
+        super();
+        writer = MAPPER.writerFor( new TypeReference<List<ApiTokenAttribute>>()
+        {
+        } );
+        reader = MAPPER.readerFor( new TypeReference<List<ApiTokenAttribute>>()
+        {
+        } );
+        returnedClass = ApiTokenAttribute.class;
+    }
+
+    @Override
+    protected void init( Class<?> klass )
+    {
+        returnedClass = klass;
+        writer = MAPPER.writerFor( new TypeReference<List<ApiTokenAttribute>>()
+        {
+        } );
+        reader = MAPPER.readerFor( new TypeReference<List<ApiTokenAttribute>>()
+        {
+        } );
+    }
+}
